backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
typing_extensions==4.12.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
